requirements.txt
openai==1.12.0
python-dotenv==1.0.0
Pillow==10.2.0
youtube-transcript-api==0.6.2
google-generativeai==0.3.2 